Migrate PantsProjectCache to a service
