+++ resources/META-INF/plugin.xml
@@ -37,10 +37,6 @@
       <interface-class>com.twitter.intellij.pants.components.PantsProjectComponent</interface-class>
       <implementation-class>com.twitter.intellij.pants.components.impl.PantsProjectComponentImpl</implementation-class>
     </component>
-    <component>
-      <interface-class>com.twitter.intellij.pants.components.PantsProjectCache</interface-class>
-      <implementation-class>com.twitter.intellij.pants.components.impl.PantsProjectCacheImpl</implementation-class>
-    </component>
   </project-components>
 
   <actions>
@@ -167,6 +163,7 @@
                 conditionClass="com.twitter.intellij.pants.ui.PantsToolWindowFactoryCondition"/>
     <externalSystemNotificationExtension implementation="com.twitter.intellij.pants.notification.PantsNotificationCustomizer"/>
     <externalProjectDataService implementation="com.twitter.intellij.pants.service.project.metadata.PantsMetadataService"/>
+    <projectService serviceImplementation="com.twitter.intellij.pants.components.PantsProjectCache"/>
   </extensions>
   <projectListeners>
     <listener class="com.twitter.intellij.pants.execution.SetJarManifestCommandLineShortener"
